reducible component javadoc
