--- src/main/java/com/github/thedeathlycow/thermoo/api/environment/component/ReducibleComponent.java
@@ -1,8 +1,0 @@
-package com.github.thedeathlycow.thermoo.api.environment.component;
-
-import org.jetbrains.annotations.Contract;
-
-public interface ReducibleComponent<T> {
-    @Contract("_->new")
-    T mergeWith(T other);
-}
